backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-jose>=3.3.0
passlib>=1.7.4
pathway>=0.14.0
